# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Category:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class:
    properties:
      capacity:
        type: integer
      category:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Category'
      category_id:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        items:
          type: integer
        type: array
      id:
        type: string
      name:
        type: string
      price:
        description: 'Financial Data: Always use int64 for money, price, balance,
          or amounts'
        type: integer
      tenant_id:
        type: string
      type:
        description: '''private'' or ''group'''
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule:
    properties:
      class:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class'
      class_id:
        type: string
      day_of_week:
        description: 1 = Senin, 7 = Minggu (ISO 8601)
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      end_time:
        description: standard HH:MM:SS for time column
        type: string
      enrollment:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment'
      enrollment_id:
        type: string
      id:
        type: string
      location:
        type: string
      start_time:
        description: standard HH:MM:SS for time column
        type: string
      tutor_id:
        type: string
      valid_from:
        type: string
      valid_until:
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSession:
    properties:
      class:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class'
      class_id:
        type: string
      end_time:
        type: string
      enrollment:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment'
      enrollment_id:
        type: string
      id:
        type: string
      schedule:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSchedule'
      schedule_id:
        type: string
      session_date:
        type: string
      start_time:
        type: string
      status:
        description: '''scheduled'', ''rescheduled'', ''cancelled'', ''completed'''
        type: string
      tutor_id:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateStudentRequest:
    properties:
      date_of_birth:
        type: string
      full_name:
        type: string
      parent_id:
        type: string
    required:
    - date_of_birth
    - full_name
    - parent_id
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Enrollment:
    properties:
      class:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Class'
      class_id:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      joined_at:
        type: string
      status:
        description: '''pending'', ''active'', ''completed'', ''dropped'''
        type: string
      student:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Student'
      student_id:
        type: string
      tenant_id:
        description: Cross-service, ordinary UUID
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse:
    properties:
      data: {}
      message:
        example: An error occurred
        type: string
      status:
        example: error
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      status:
        example: success
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.SessionListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSession'
        type: array
      pagination:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Pagination'
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Student:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      full_name:
        type: string
      id:
        type: string
      parent_id:
        description: Cross-service, ordinary UUID
        type: string
      updated_at:
        type: string
    type: object
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.StudentListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Student'
        type: array
      pagination:
        $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Pagination'
    type: object
...
  github_com_kelolakelas_kelolakelas-academic-service_internal_domain.UpdateStudentRequest:
    properties:
      date_of_birth:
        type: string
      full_name:
        type: string
    required:
    - date_of_birth
    - full_name
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
  description: Academic Management Service for KelolaKelas Platform
  title: KelolaKelas Academic Service API
  version: "1.0"
paths:
...
  /api/v1/sessions:
    get:
      description: List class sessions scoped to the active tenant
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.SessionListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List tenant sessions
      tags:
      - Sessions
  /api/v1/sessions/{id}:
    get:
      description: Get one session scoped to the active tenant
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ClassSession'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tenant session
      tags:
      - Sessions
...
  /api/v1/students:
    get:
      description: List students scoped to the authenticated parent or tenant enrollments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.StudentListResponse'
              type: object
      security:
      - BearerAuth: []
      summary: List students
      tags:
      - Students
    post:
      consumes:
      - application/json
      parameters:
      - description: Student payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.CreateStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Student'
              type: object
      security:
      - BearerAuth: []
      summary: Create student
      tags:
      - Students
  /api/v1/students/{id}:
    delete:
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete student
      tags:
      - Students
    get:
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Student'
              type: object
      security:
      - BearerAuth: []
      summary: Get student
      tags:
      - Students
    patch:
      consumes:
      - application/json
      parameters:
      - description: Student UUID
        in: path
        name: id
        required: true
        type: string
      - description: Student payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.UpdateStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.HTTPResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_kelolakelas_kelolakelas-academic-service_internal_domain.Student'
              type: object
      security:
      - BearerAuth: []
      summary: Update student
      tags:
      - Students
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"